refactor(frontend): extract savedScenarioToScenario to its own module
Move the pure SavedScenario -> Scenario transform out of ScenarioContext.tsx
and into frontend/src/contexts/scenarioTransform.ts. This removes the
react-refresh/only-export-components eslint-disable workaround: the context
file no longer exports non-component values alongside the provider, and the
transform is directly testable without importing from a component module.

Existing unit tests move from useSavedScenariosMutation.test.ts into a new
scenarioTransform.test.ts alongside the new module. Behavior is unchanged.
